:hammer: chore extract create-record function
